backend/requirements.txt
# FastAPI Backend Dependencies for Sony Interior

# Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.34.0

# AI - Google Gemini
google-genai

# Database
asyncpg>=0.30.0

# Data Validation
pydantic>=2.10.0
pydantic-settings

# Environment Variables
python-dotenv

# HTTP Client
httpx>=0.28.0

# Form Data
python-multipart

# CORS and Security
python-jose[cryptography]
passlib[bcrypt]

# Utilities
python-dateutil

# HTTP Server for MCP
aiohttp
